* add dialog to edit water amount * added settings screen * added notification icon

+++ app/src/main/java/com/criticalay/neer/data/model/Beverage.kt
@@ -28,7 +28,7 @@ import com.criticalay.neer.utils.Constants.BEVERAGE_DATABASE_TABLE
 data class Beverage(
     val userId:Long,
     val beverageName: String,
-    val totalIntakeAmount : Double,
+    val totalIntakeAmount : Int,
     @PrimaryKey
     val beverageId: Long = 101,
 )
